Add logic to delete files while deleting account in delete-user-form.blade.php component

+++ resources/views/livewire/settings/delete-user-form.blade.php
@@ -29,11 +29,23 @@ new class extends Component {
 
         //tap(Auth::user(), $logout(...))->delete();
 
-        Storage::disk('avatars')->delete(Auth::user()->avatar);
+        if(Auth::user()->avatar)
+        {
+            Storage::disk('avatars')->delete(Auth::user()->avatar);
+        }
+
+        //delete files of all the announcements of the user
+        foreach(Auth::user()->senderAnnouncements as $senderannouncement)
+        {
+            foreach($senderannouncement->library as $image)
+            {
+                Storage::disk('senders-announcements')->delete($image['path']);
+            }
+        }
 
         Auth::user()->delete();
 
-        Auth::guard('web')->logout();
+        //Auth::guard('web')->logout();
  
         Session::invalidate();
